requirements.txt
fastapi
pillow
python-multipart
pyopenssl
pdfrw